validator/requirements.txt
pandas==2.2.2
sweetviz
dtale
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
clickhouse-driver==0.2.9
great_expectations==0.17.22
openlineage-python==1.19.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
mlflow
scikit-learn
prometheus-client 
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
clickhouse-connect==0.7.0
python-multipart==0.0.6
aiofiles==23.2.1
